Cache missing information, to prevent inadvertent performance issues
This should not happen for production, but I think it's better to
cache the misses in case someone forgets creating the release
version/date files.

The only downside I see is that missing files can't be added after the
server was started and the first request was made.

Also remove an unused method and several unused files.